+++ src/lua-ssi-header.lua
@@ -1,5 +1,16 @@
 if not ngx.headers_sent and ngx.ctx.res
 then
+    local requestHeaders = {}
+
+    local sanatizeHeaderFieldName = function(headerFieldName)
+        return string.gsub(string.lower(headerFieldName), "_", "-")
+    end
+
+    for k, v in pairs(ngx.req.get_headers()) do
+        k = sanatizeHeaderFieldName(k)
+        requestHeaders[k] = v
+    end
+
     for k, v in pairs(ngx.ctx.res.header) do
         ngx.header[k] = v
     end
@@ -11,6 +22,11 @@ then
     end
     if ngx.ctx.ssiDepth then
         ngx.header["X-Ssi-Depth"] = ngx.ctx.ssiDepth
+    end
+    if requestHeaders["x-ssi-debug"] == "true" then
+        ngx.header["X-Ssi-Minimize-MaxAge-Url"] = ngx.ctx.ssiMinimizeMaxAgeUrl
+        ngx.header["X-Ssi-Minimize-MaxAge-Age"] = ngx.ctx.ssiMinimizeMaxAgeAge
+        ngx.header["X-Ssi-Minimize-MaxAge-Cache-Control"] = ngx.ctx.ssiMinimizeMaxAgeCacheControl
     end
     if ngx.ctx.overrideContentType then
         ngx.header["Content-Type"] = ngx.ctx.overrideContentType